Prevent get_armed_area_state() overriding part armed to fully armed
When get_armed_area_state() polls (regularly) for area flags, we get a
binary disarmed / armed flag. If the area is part armed, the "armed"
flag is 1, but currently this will override the "part armed" current
state of the area and move it to "fully armed".
Prevent this by checking if the current area state is part armed, and
the new state would be fully armed, and prevent this override if so.

Signed-off-by: Matt Redfearn <matt.redfearn@gmail.com>